+++ Types.h
@@ -12,6 +12,6 @@ enum class Side {
 struct Order {
     uint64_t order_id; //Massive size for order id
     Side side;
-    double price;
-    uint32_t quantity;
+    uint64_t price;
+    uint64_t quantity;
 };
